# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Work Order Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  student: "student",
  faculty: "faculty",
  admin: "admin",
  technician: "technician",
} as const;

export interface RegisterRequest {
  email: string;
  password: string;
  full_name: string;
  role: RegisterRequestRole;
  department?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  faculty: "faculty",
  admin: "admin",
  technician: "technician",
} as const;

export interface User {
  id: string;
  email: string;
  full_name: string;
  role: UserRole;
  department?: string;
  created_at?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type WorkOrderCategory =
  (typeof WorkOrderCategory)[keyof typeof WorkOrderCategory];

export const WorkOrderCategory = {
  electrical: "electrical",
  plumbing: "plumbing",
  it: "it",
  lab_equipment: "lab_equipment",
  safety: "safety",
  general: "general",
} as const;

export type WorkOrderPriority =
  (typeof WorkOrderPriority)[keyof typeof WorkOrderPriority];

export const WorkOrderPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export type WorkOrderStatus =
  (typeof WorkOrderStatus)[keyof typeof WorkOrderStatus];

export const WorkOrderStatus = {
  open: "open",
  assigned: "assigned",
  in_progress: "in_progress",
  complete: "complete",
  closed: "closed",
} as const;

export interface WorkOrder {
  id: string;
  title: string;
  description: string;
  category: WorkOrderCategory;
  location: string;
  building?: string;
  priority: WorkOrderPriority;
  status: WorkOrderStatus;
  requester_id: string;
  requester_name?: string;
  assigned_to_id?: string;
  assigned_to_name?: string;
  rejection_reason?: string;
  created_at: string;
  updated_at?: string;
}

export type CreateWorkOrderRequestCategory =
  (typeof CreateWorkOrderRequestCategory)[keyof typeof CreateWorkOrderRequestCategory];

export const CreateWorkOrderRequestCategory = {
  electrical: "electrical",
  plumbing: "plumbing",
  it: "it",
  lab_equipment: "lab_equipment",
  safety: "safety",
  general: "general",
} as const;

export type CreateWorkOrderRequestPriority =
  (typeof CreateWorkOrderRequestPriority)[keyof typeof CreateWorkOrderRequestPriority];

export const CreateWorkOrderRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface CreateWorkOrderRequest {
  title: string;
  description: string;
  category: CreateWorkOrderRequestCategory;
  location: string;
  building?: string;
  priority: CreateWorkOrderRequestPriority;
}

export interface UpdateWorkOrderRequest {
  title?: string;
  description?: string;
  category?: string;
  location?: string;
  building?: string;
  priority?: string;
}

export interface RejectRequest {
  reason: string;
}

export interface AssignRequest {
  technician_id: string;
}

export type StatusUpdateRequestStatus =
  (typeof StatusUpdateRequestStatus)[keyof typeof StatusUpdateRequestStatus];

export const StatusUpdateRequestStatus = {
  open: "open",
  assigned: "assigned",
  in_progress: "in_progress",
  complete: "complete",
  closed: "closed",
} as const;

export interface StatusUpdateRequest {
  status: StatusUpdateRequestStatus;
  notes?: string;
}

export interface Comment {
  id: string;
  work_order_id: string;
  author_id: string;
  author_name?: string;
  content: string;
  created_at: string;
}

export interface AddCommentRequest {
  content: string;
}

export interface Notification {
  id: string;
  user_id: string;
  message: string;
  type: string;
  work_order_id?: string;
  is_read: boolean;
  created_at: string;
}

export type DashboardStatsByCategory = { [key: string]: number };

export interface DashboardStats {
  total: number;
  open: number;
  assigned: number;
  in_progress: number;
  complete: number;
  closed: number;
  urgent: number;
  by_category?: DashboardStatsByCategory;
}

export type ListUsersParams = {
  role?: ListUsersRole;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  student: "student",
  faculty: "faculty",
  admin: "admin",
  technician: "technician",
} as const;

export type ListWorkOrdersParams = {
  status?: ListWorkOrdersStatus;
  priority?: ListWorkOrdersPriority;
  category?: string;
};

export type ListWorkOrdersStatus =
  (typeof ListWorkOrdersStatus)[keyof typeof ListWorkOrdersStatus];

export const ListWorkOrdersStatus = {
  open: "open",
  assigned: "assigned",
  in_progress: "in_progress",
  complete: "complete",
  closed: "closed",
} as const;

export type ListWorkOrdersPriority =
  (typeof ListWorkOrdersPriority)[keyof typeof ListWorkOrdersPriority];

export const ListWorkOrdersPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;