# lib/api-zod/src/generated/types/addCommentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Work Order Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AddCommentRequest {
  content: string;
}