# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Work Order Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: string;
  work_order_id: string;
  author_id: string;
  author_name?: string;
  content: string;
  created_at: string;
}