# lib/api-zod/src/generated/types/createWorkOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Work Order Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateWorkOrderRequestCategory } from "./createWorkOrderRequestCategory";
import type { CreateWorkOrderRequestPriority } from "./createWorkOrderRequestPriority";

export interface CreateWorkOrderRequest {
  title: string;
  description: string;
  category: CreateWorkOrderRequestCategory;
  location: string;
  building?: string;
  priority: CreateWorkOrderRequestPriority;
}