# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Work Order Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsByCategory } from "./dashboardStatsByCategory";

export interface DashboardStats {
  total: number;
  open: number;
  assigned: number;
  in_progress: number;
  complete: number;
  closed: number;
  urgent: number;
  by_category?: DashboardStatsByCategory;
}